refactors and bug fixes
